pyproject.toml
[project]
name = "macro-insight-engine"
version = "0.1.0"
description = "Automated pipeline to synthesize global macro sentiment from financial podcasts."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.47.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "ruff>=0.15.0",
    "yt-dlp>=2025.10.14",
]

[project.scripts]
mie = "mie.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mie"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

